Merge development into master

+++ script/interactiveDeployment.sh
@@ -39,6 +39,7 @@ fi
 
 eulerEarnCompilerOptions="--via-ir --optimize --optimizer-runs 200 --use 0.8.26 --out out-euler-earn"
 eulerSwapCompilerOptions="--optimize --optimizer-runs 2500 --use 0.8.27 --out out-euler-swap"
+securitizeFactoryCompilerOptions="--optimize --optimizer-runs 10000 --use 0.8.24 --out out-securitize-factory"
 
 while true; do
     echo ""
@@ -60,8 +61,7 @@ while true; do
     echo "14. Bridging contracts"
     echo "15. Edge factory"
     echo "---------------------------------"
-    echo "20. EulerEarn implementation (modules and implementation contract)"
-    echo "21. EulerEarn factory"
+    echo "20. EulerEarn factory and public allocator"
     echo "---------------------------------"
     echo "50. Core and Periphery Deployment and Configuration"
     echo "51. Core Ownership Transfer"
@@ -561,7 +561,7 @@ while true; do
             echo "Select the type of IRM to deploy:"
             echo "0. Kink"
             echo "1. Adaptive Curve"
-            read -p "Enter your choice (0-0): " irm_choice
+            read -p "Enter your choice (0-1): " irm_choice
 
             baseName=04_IRM
 
@@ -774,6 +774,12 @@ while true; do
                     read -p "Enter the EVault Factory address: " eVaultFactory
                     read -p "Enter the Oracle Adapter Registry address: " oracle_adapter_registry
                     read -p "Enter the Kink IRM Factory address: " kink_irm_factory
+                    read -p "Enter the Adaptive Curve IRM Factory address (or press Enter for address(0)): " adaptive_curve_irm_factory
+                    adaptive_curve_irm_factory=${adaptive_curve_irm_factory:-0x0000000000000000000000000000000000000000}
+                    read -p "Enter the Kinky IRM Factory address (or press Enter for address(0)): " kinky_irm_factory
+                    kinky_irm_factory=${kinky_irm_factory:-0x0000000000000000000000000000000000000000}
+                    read -p "Enter the Fixed Cyclical Binary IRM Factory address (or press Enter for address(0)): " fixed_cyclical_binary_irm_factory
+                    fixed_cyclical_binary_irm_factory=${fixed_cyclical_binary_irm_factory:-0x0000000000000000000000000000000000000000}
 
                     jq -n \
                         --arg eVaultFactory "$eVaultFactory" \
@@ -1216,7 +1222,7 @@ while true; do
             jq -n \
                 --arg evc "$evc" \
                 --arg permit2 "$permit2" \
-                --arg evc "$perspective" \
+                --arg perspective "$perspective" \
                 '{
                     evc: $evc,
                     permit2: $permit2,
@@ -1243,10 +1249,11 @@ while true; do
                 multisig_security_partner_B=$(jq -r '.securityPartnerB' "$addresses_dir_path/MultisigAddresses.json" 2>/dev/null)
                 evc=$(jq -r '.evc' "$addresses_dir_path/CoreAddresses.json" 2>/dev/null)
                 swapper=$(jq -r '.swapper' "$addresses_dir_path/PeripheryAddresses.json" 2>/dev/null)
+                feeFlowController=$(jq -r '.feeFlowController' "$addresses_dir_path/PeripheryAddresses.json" 2>/dev/null)
+                securitizeFactory=$(jq -r '.securitizeFactory' "$addresses_dir_path/PeripheryAddresses.json" 2>/dev/null)
                 eulOFTAdapter=$(jq -r '.eulOFTAdapter' "$addresses_dir_path/BridgeAddresses.json" 2>/dev/null)
                 eusdOFTAdapter=$(jq -r '.eusdOFTAdapter' "$addresses_dir_path/BridgeAddresses.json" 2>/dev/null)
                 seusdOFTAdapter=$(jq -r '.seusdOFTAdapter' "$addresses_dir_path/BridgeAddresses.json" 2>/dev/null)
-                feeFlowController=$(jq -r '.feeFlowController' "$addresses_dir_path/PeripheryAddresses.json" 2>/dev/null)
                 eulerEarnFactory=$(jq -r '.eulerEarnFactory' "$addresses_dir_path/CoreAddresses.json" 2>/dev/null)
                 eulerEarnFactory=${eulerEarnFactory:-$addressZero}
                 eulerSwapV2Factory=$(jq -r '.eulerSwapV2Factory' "$addresses_dir_path/EulerSwapAddresses.json" 2>/dev/null)
@@ -1299,6 +1306,10 @@ while true; do
                 read -p "Should deploy and configure seUSD contracts system? (y/n) (default: n): " deploy_seusd
             fi
 
+            if [ -z "$securitizeFactory" ] || [ "$securitizeFactory" == "$addressZero" ] || [ "$securitizeFactory" == "null" ]; then
+                read -p "Should deploy Securitize Vault Factory? (y/n) (default: n): " deploy_securitize_factory
+            fi
+
             multisig_dao=${multisig_dao:-$addressZero}
             multisig_labs=${multisig_labs:-$addressZero}
             multisig_security_council=${multisig_security_council:-$addressZero}
@@ -1313,6 +1324,7 @@ while true; do
             deploy_euler_swap=${deploy_euler_swap:-n}
             deploy_eusd=${deploy_eusd:-n}
             deploy_seusd=${deploy_seusd:-n}
+            deploy_securitize_factory=${deploy_securitize_factory:-n}
             uniswap_pool_manager=${uniswap_pool_manager:-$addressZero}
             euler_swap_protocol_fee_config_admin=${euler_swap_protocol_fee_config_admin:-$multisig_dao}
             euler_swap_registry_curator=${euler_swap_registry_curator:-$multisig_labs}
@@ -1323,6 +1335,10 @@ while true; do
 
             if { [ -z "$eulerSwapV2Factory" ] || [ "$eulerSwapV2Factory" == "$addressZero" ] || [ "$eulerSwapV2Factory" == "null" ]; } && [ "$deploy_euler_swap" = "y" ]; then
                 forge compile lib/euler-swap/src $eulerSwapCompilerOptions --force
+            fi
+
+            if { [ -z "$securitizeFactory" ] || [ "$securitizeFactory" == "$addressZero" ] || [ "$securitizeFactory" == "null" ]; } && [ "$deploy_securitize_factory" = "y" ]; then
+                forge compile src/VaultFactory/ERC4626EVCCollateralSecuritizeFactory.sol $securitizeFactoryCompilerOptions --force
             fi
 
             if [[ "$@" != *"--ffi"* ]]; then
@@ -1344,6 +1360,7 @@ while true; do
                 --argjson deployEulerSwap "$(jq -n --argjson val \"$deploy_euler_swap\" 'if $val == "y" then true else false end')" \
                 --argjson deployEUSD "$(jq -n --argjson val \"$deploy_eusd\" 'if $val == "y" then true else false end')" \
                 --argjson deploySEUSD "$(jq -n --argjson val \"$deploy_seusd\" 'if $val == "y" then true else false end')" \
+                --argjson deploySecuritizeFactory "$(jq -n --argjson val \"$deploy_securitize_factory\" 'if $val == "y" then true else false end')" \
                 --arg uniswapPoolManager "$uniswap_pool_manager" \
                 --arg eulerSwapProtocolFeeConfigAdmin "$euler_swap_protocol_fee_config_admin" \
                 --arg eulerSwapRegistryCurator "$euler_swap_registry_curator" \
@@ -1362,6 +1379,7 @@ while true; do
                     deployEulerSwap: $deployEulerSwap,
                     deployEUSD: $deployEUSD,
                     deploySEUSD: $deploySEUSD,
+                    deploySecuritizeFactory: $deploySecuritizeFactory,
                     uniswapPoolManager: $uniswapPoolManager,
                     eulerSwapProtocolFeeConfigAdmin: $eulerSwapProtocolFeeConfigAdmin,
                     eulerSwapRegistryCurator: $eulerSwapRegistryCurator
@@ -1439,7 +1457,7 @@ while true; do
                     ;;
             esac
 
-            cast send $access_control_contract_address $signature $bytes32_role_identifier $account_address --rpc-url $DEPLOYMENT_RPC_URL --legacy $broadcast $@
+            cast send $access_control_contract_address $signature $bytes32_role_identifier $account_address --rpc-url $DEPLOYMENT_RPC_URL --legacy $@
             ;;
         *)
             echo "Invalid choice. Exiting."
@@ -1470,6 +1488,7 @@ while true; do
         cp "$broadcast_dir/run-latest.json" "$deployment_dir/broadcast/${jsonName}_${counter}.json"
 
         for json_file in script/*_input.json; do
+            [ -e "$json_file" ] || continue
             jsonFileName=$(basename "${json_file/_input/}")
             counter=$(script/utils/getFileNameCounter.sh "$deployment_dir/input/$jsonFileName")
 
@@ -1477,6 +1496,7 @@ while true; do
         done
 
         for json_file in script/*_output.json; do
+            [ -e "$json_file" ] || continue
             jsonFileName=$(basename "${json_file/_output/}")
             counter=$(script/utils/getFileNameCounter.sh "$deployment_dir/output/$jsonFileName")
 
@@ -1484,6 +1504,7 @@ while true; do
         done
 
         for json_file in script/*.json; do
+            [ -e "$json_file" ] || continue
             jsonFileName=$(basename "${json_file}")
             counter=$(script/utils/getFileNameCounter.sh "$deployment_dir/output/$jsonFileName")
 
@@ -1491,6 +1512,7 @@ while true; do
         done
     else
         for json_file in script/*.json; do
+            [ -e "$json_file" ] || continue
             rm "$json_file"
         done
     fi
